Move Switch component to src/components/elementsList folder
