+++ app/packs/src/components/elementsList/Switch.js
